Extract TestDataGen class into a helpers subpackage
* extracted the TestDataGen class in user_test.py into a new test
helpers subpackage in a module called testdata_gen.py so that it can be
re-used and extended in other tests
